Fix some missing QMK_KEYBOARD_H includes in user keymaps
