+++ keyboards/1upkeyboards/sweet16/keymaps/dlaroe/keymap.c
@@ -1,4 +1,4 @@
-#include "sweet16.h"
+#include QMK_KEYBOARD_H
 
 enum layers {
  num,
