+++ keyboards/bpiphany/hid_liber/keymaps/bakageta/keymap.c
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "hid_liber.h"
+#include QMK_KEYBOARD_H
 
 // Helpful defines
 #define FN_CAPS LT(_FL, KC_CAPS)
