+++ keyboards/contra/keymaps/maxr1998/keymap.c
@@ -14,7 +14,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "contra.h"
+#include QMK_KEYBOARD_H
 #include "keymap_german.h"
 
 enum contra_layers {
