+++ keyboards/contra/keymaps/ryanm101/keymap.c
@@ -1,4 +1,4 @@
-#include "contra.h"
+#include QMK_KEYBOARD_H
 #include "action_layer.h"
 
 extern keymap_config_t keymap_config;
