+++ keyboards/deltasplit75/keymaps/mbsurfer/keymap.c
@@ -1,6 +1,4 @@
-#include "deltasplit75.h"
-#include "action_layer.h"
-#include "eeconfig.h"
+#include QMK_KEYBOARD_H
 
 extern keymap_config_t keymap_config;
