+++ keyboards/deltasplit75/keymaps/protosplit/config.h
@@ -19,11 +19,3 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define MASTER_LEFT
 // #define MASTER_RIGHT
 // #define EE_HANDS
-
-
-#ifdef SUBPROJECT_v2
-    #include "../../v2/config.h"
-#endif
-#ifdef SUBPROJECT_protosplit
-    #include "../../protosplit/config.h"
-#endif
