+++ keyboards/ergodox_ez/keymaps/rmw/keymap.c
@@ -1,4 +1,4 @@
-#include "ergodox_ez.h"
+#include QMK_KEYBOARD_H
 #include "version.h"
 
 #define LONGPRESS_DELAY 180
