+++ keyboards/handwired/xealous/rules.mk
@@ -11,7 +11,6 @@ BACKLIGHT_ENABLE = no       # Enable keyboard backlight functionality
 AUDIO_ENABLE = yes          # Audio output
 RGBLIGHT_ENABLE = no        # Enable WS2812 RGB underlight.
 SPLIT_KEYBOARD = yes        # Use shared split_common code
-SUBPROJECT_rev1 = yes
 
 SRC += matrix.c
 DEBOUNCE_TYPE = sym_eager_pk
