+++ keyboards/keebio/chocopad/keymaps/khord/keymap.c
@@ -1,4 +1,4 @@
-#include "chocopad.h"
+#include QMK_KEYBOARD_H
 
 #define _BASE 0
 #define _FN1 1
