+++ keyboards/lets_split/keymaps/heartrobotninja/config.h
@@ -41,14 +41,4 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #undef TAPPING_TOGGLE
 #define TAPPING_TOGGLE 2
 
-#ifdef SUBPROJECT_rev1
-#include "../../rev1/config.h"
-#endif
-#ifdef SUBPROJECT_rev2
-#include "../../rev2/config.h"
-#endif
-#ifdef SUBPROJECT_rev2fliphalf
-#include "../../rev2fliphalf/config.h"
-#endif
-
 #endif
