+++ keyboards/lets_split/keymaps/piemod/rules.mk
@@ -10,4 +10,3 @@ AUDIO_ENABLE = no           # Audio output on port C6
 UNICODE_ENABLE = yes         # Unicode
 BLUETOOTH_ENABLE = no       # Enable Bluetooth with the Adafruit EZ-Key HID
 RGBLIGHT_ENABLE = yes       # Enable WS2812 RGB underlight. 
-SUBPROJECT_rev1 = no
