+++ keyboards/mehkee96/keymaps/johann/keymap.c
@@ -1,4 +1,4 @@
-#include "mehkee96.h"
+#include QMK_KEYBOARD_H
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
