+++ keyboards/orthodox/keymaps/oscillope/rules.mk
@@ -1,3 +1,2 @@
-SUBPROJECT_rev1 = no
 COMMAND_ENABLE = no
 MOUSEKEY_ENABLE = no
