+++ keyboards/planck/keymaps/ishtob/keymap.c
@@ -1,13 +1,6 @@
 // This is the canonical layout file for the Quantum project. If you want to add another keyboard,
 // this is the style you want to emulate.
 
-#include "planck.h"
-#include "action_layer.h"
-// #include "dynamic_macro.h"
-#ifdef AUDIO_ENABLE
-  #include "audio.h"
-#endif
-#include "eeconfig.h"
 #include "ishtob.h"
 
 extern keymap_config_t keymap_config;
