--- keyboards/planck/keymaps/mitch/rules.mk
@@ -1,2 +1,0 @@
-SUBPROJECT = rev3
-
