+++ keyboards/planck/keymaps/piemod/rules.mk
@@ -1,5 +1,3 @@
-SUBPROJECT_DEFAULT = rev4
-
 MOUSEKEY_ENABLE = yes
 EXTRAKEY_ENABLE = yes
 AUDIO_ENABLE = yes
