+++ keyboards/woodkeys/meira/keymaps/grahampheath/keymap.c
@@ -13,9 +13,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "meira.h"
-#include "action_layer.h"
-#include "action_code.h"
+#include QMK_KEYBOARD_H
 
 #ifdef RGBLIGHT_ENABLE
 //Following line allows macro to read current RGB settings
