+++ keyboards/ymdk/ymd96/keymaps/hgoel89/keymap.c
@@ -1,5 +1,4 @@
-#include "ymd96.h"
-#include "action_layer.h"
+#include QMK_KEYBOARD_H
 
 /*
 #define _QWERTY 0
